refactor: api constants
