+++ src/api/axios.ts
@@ -1,6 +1,6 @@
 import type { InternalAxiosRequestConfig } from 'axios';
 import axios from 'axios';
-import { BASE_URL } from './constants';
+import { BASE_URL } from '../sources/constants/api';
 
 export const api = axios.create({
   baseURL: BASE_URL,
